requirements.txt
anthropic>=0.44.0
playwright>=1.49.0
pyyaml>=6.0
python-dotenv>=1.0
apscheduler>=3.10.4
Pillow>=10.0.0
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
pytest>=7.0.0
